% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod_functions.R
\name{generate_lods}
\alias{generate_lods}
\title{Compute Limits of  Detection}
\usage{
generate_lods(param_ci_df, verbose = TRUE)
}
\arguments{
\item{param_ci_df}{Long-format tibble (freq_param_ci_df / bayes_param_ci_df)}

\item{verbose}{Logical. If \code{TRUE}, emit one diagnostic message per curve.}
}
\value{
Tibble: curve_id, llod, ulod'
}
\description{
Compute Limits of  Detection
}
